+++ example/lib/homepage/network_tools/bili_favorite_video/bili_favorite_video_logic.dart
@@ -30,7 +30,7 @@ class BiliFavoriteVideoLogic extends PagingController2<BiliFavoriteVideo, BiliFa
   }
 
   @override
-  BiliFavoriteVideoPagingState getPagingState() {
+  BiliFavoriteVideoPagingState customPagingState() {
     return state.biliFavoriteVideoPagingState;
   }
 }