+++ example/lib/homepage/network_tools/bili_hot_video/bili_hot_video_logic.dart
@@ -36,7 +36,7 @@ class BiliHotVideoLogic extends PagingController<BiliHotVideo, BiliHotVideoPagin
   }
 
   @override
-  BiliHotVideoPagingState providePagingState() {
+  BiliHotVideoPagingState customPagingState() {
     return state.biliHotVideoPagingState;
   }